put registry auth
